core: make holiday-calendar coverage explicit and extend it to 2000-2035
is_holiday was a bare map lookup over a calendar spanning only 2025-2027, so
any date outside that range returned false -- indistinguishable from 'the market
was open'. A replay or backtest of an earlier year therefore walked
find_previous_trading_day onto a closed day, loaded an empty previous-day book,
and silently skipped corporate actions while mis-stating PnL, with no symptom.
The calendar also ran out after 2027, so the same failure was scheduled to
arrive on its own.

HolidayChecker now records which years it loaded and exposes covers_date /
covers_year / coverage_years / coverage_description, warns once per out-of-range
year rather than answering silently, and the equity runner fails closed when the
calendar cannot cover the date it is about to trade.

The calendar itself is now generated by scripts/generate_market_holidays.py for
2000-2035 from the NYSE rules (fixed and nth-weekday holidays, Good Friday via
the Gregorian computus, Saturday-observed-Friday / Sunday-observed-Monday, the
New Year's Saturday exception where no closure is observed, and Juneteenth only
from 2022). The generator reproduces all 30 previously hand-maintained 2025-2027
entries exactly, which is what validates its rules; --check re-runs that
comparison. Unscheduled closures are not derivable and are carried explicitly.

Two deltas to dates that were already covered, both corrections rather than
additions: 2025-01-09 (National Day of Mourning, Carter) was missing from the
shipped file, and years before 2025 are now populated. Both change what
is_holiday returns for replays of those dates on BOTH asset classes; the futures
integration baseline (2026-05-02) is unaffected, and no futures source file was
touched.

Docs corrected: the class described itself as a federal holiday checker while
shipping the market calendar, which includes Good Friday and excludes Columbus
Day and Veterans Day.
