+++ tests/backtest/test_backtest_metrics_calculator.cpp
