+++ tests/backtest/test_backtest_price_manager.cpp
