--- tests/backtesting/test_engine.cpp
@@ -1,875 +1,0 @@
-#include <gtest/gtest.h>
-#include <algorithm>
-#include <chrono>
-#include <memory>
-#include <numeric>
-#include <random>
-#include <thread>
-#include "../data/test_db_utils.hpp"
-#include "trade_ngin/backtest/backtest_engine.hpp"
-#include "trade_ngin/core/state_manager.hpp"
-#include "trade_ngin/strategy/base_strategy.hpp"
-
-using namespace trade_ngin;
-using namespace trade_ngin::backtest;
-using namespace trade_ngin::testing;
-
-class MockStrategy : public BaseStrategy {
-public:
-    MockStrategy()
-        : BaseStrategy(
-              "mock_strategy",
-              []() {
-                  StrategyConfig config;
-                  config.capital_allocation = 100000.0;
-                  config.max_leverage = 20.0;
-                  config.position_limits = {{"AAPL", 1000.0}};
-                  config.max_drawdown = 0.5;
-                  config.var_limit = 0.25;
-                  config.correlation_limit = 0.5;
-                  config.trading_params = {{"AAPL", 1.0}};
-                  config.costs = {{"AAPL", 0.005}};
-                  config.asset_classes = {AssetClass::EQUITIES};
-                  config.frequencies = {DataFrequency::DAILY};
-                  return config;
-              }(),
-              std::make_shared<MockPostgresDatabase>("postgresql://localhost:5432/test_db")),
-          initial_position_size_(100.0) {
-        // Initialize metadata
-        metadata_.name = "Mock Strategy";
-        metadata_.description = "Mock strategy for testing";
-        metadata_.sharpe_ratio = 1.5;
-        metadata_.sortino_ratio = 1.2;
-        metadata_.max_drawdown = 0.2;
-        metadata_.win_rate = 0.6;
-
-        // Initialize with some positions to avoid "No positions provided" error
-        Position initial_pos;
-        initial_pos.symbol = "AAPL";
-        initial_pos.quantity = initial_position_size_;
-        initial_pos.average_price = 150.0;
-        initial_pos.unrealized_pnl = 0.0;
-        initial_pos.realized_pnl = 0.0;
-        initial_pos.last_update = std::chrono::system_clock::now();
-        positions_["AAPL"] = initial_pos;
-
-        // Add more initial positions for other common test symbols
-        std::vector<std::string> common_symbols = {"MSFT", "GOOG"};
-        for (const auto& symbol : common_symbols) {
-            Position pos;
-            pos.symbol = symbol;
-            pos.quantity = initial_position_size_;
-            pos.average_price = 200.0;
-            pos.unrealized_pnl = 0.0;
-            pos.realized_pnl = 0.0;
-            pos.last_update = std::chrono::system_clock::now();
-            positions_[symbol] = pos;
-        }
-    }
-
-    // Data processing with customizable behavior
-    Result<void> on_data(const std::vector<Bar>& data) override {
-        if (fail_on_data_) {
-            return make_error<void>(ErrorCode::STRATEGY_ERROR, "Simulated data failure");
-        }
-
-        // Call base class implementation
-        auto base_result = BaseStrategy::on_data(data);
-        if (base_result.is_error()) {
-            return base_result;
-        }
-
-        bars_received_ += data.size();
-
-        // Simple strategy logic: buy when price moves up, sell when it moves down
-        for (const auto& bar : data) {
-            // Get previous bar info
-            auto it = last_prices_.find(bar.symbol);
-            if (it != last_prices_.end()) {
-                double prev_price = it->second;
-                double price_change = bar.close.as_double() - prev_price;
-
-                // Generate orders based on price movement
-                if (price_change > 0 && bar.close.as_double() > prev_price * 1.005) {
-                    // Buy signal on 0.5% increase
-                    Position pos = positions_[bar.symbol];
-                    pos.quantity += trade_size_;
-                    pos.average_price = (pos.average_price * (pos.quantity - trade_size_) +
-                                         bar.close.as_double() * trade_size_) /
-                                        pos.quantity;
-                    pos.last_update = bar.timestamp;
-                    pos.unrealized_pnl = (bar.close.as_double() - pos.average_price.as_double()) *
-                                         pos.quantity.as_double();
-
-                    // Update position
-                    update_position(bar.symbol, pos);
-                } else if (price_change < 0 && bar.close.as_double() < prev_price * 0.995) {
-                    // Sell signal on 0.5% decrease
-                    Position pos = positions_[bar.symbol];
-                    if (pos.quantity.as_double() >
-                        trade_size_) {  // Ensure we maintain some position
-                        double sold_quantity = std::min(
-                            pos.quantity.as_double() - initial_position_size_, trade_size_);
-                        pos.quantity -= sold_quantity;
-                        pos.realized_pnl +=
-                            (bar.close.as_double() - pos.average_price.as_double()) * sold_quantity;
-                        pos.last_update = bar.timestamp;
-                        pos.unrealized_pnl =
-                            (bar.close.as_double() - pos.average_price.as_double()) *
-                            pos.quantity.as_double();
-
-                        // Update position
-                        update_position(bar.symbol, pos);
-                    }
-                }
-            }
-            // Update last price
-            last_prices_[bar.symbol] = bar.close.as_double();
-        }
-
-        // Ensure we have at least one position with non-zero quantity
-        if (positions_.empty()) {
-            for (const auto& bar : data) {
-                Position pos;
-                pos.symbol = bar.symbol;
-                pos.quantity = initial_position_size_;
-                pos.average_price = bar.close.as_double();
-                pos.last_update = bar.timestamp;
-                pos.unrealized_pnl = (bar.close.as_double() - pos.average_price.as_double()) *
-                                     pos.quantity.as_double();
-                pos.realized_pnl = 0.0;
-                update_position(bar.symbol, pos);
-                break;
-            }
-        }
-
-        return Result<void>();
-    }
-
-    // Override on_execution to track executions received
-    Result<void> on_execution(const ExecutionReport& report) override {
-        executions_received_++;
-        return BaseStrategy::on_execution(report);
-    }
-
-    // Override on_signal to track signals received
-    Result<void> on_signal(const std::string& symbol, double signal) override {
-        signals_received_++;
-        return BaseStrategy::on_signal(symbol, signal);
-    }
-
-    // Test control methods
-    void set_fail_on_data(bool fail) {
-        fail_on_data_ = fail;
-    }
-    void set_trade_size(double size) {
-        trade_size_ = size;
-    }
-    void set_initial_position_size(double size) {
-        initial_position_size_ = size;
-    }
-    int get_bars_received() const {
-        return bars_received_;
-    }
-    int get_executions_received() const {
-        return executions_received_;
-    }
-    int get_signals_received() const {
-        return signals_received_;
-    }
-
-    // Add a position directly for testing
-    void add_position(const std::string& symbol, double quantity, double price) {
-        Position pos;
-        pos.symbol = symbol;
-        pos.quantity = quantity;
-        pos.average_price = price;
-        pos.unrealized_pnl = 0.0;
-        pos.realized_pnl = 0.0;
-        pos.last_update = std::chrono::system_clock::now();
-        positions_[symbol] = pos;
-    }
-
-private:
-    std::unordered_map<std::string, double> last_prices_;
-    bool fail_on_data_ = false;
-    double trade_size_ = 10.0;
-    double initial_position_size_ = 100.0;
-    int bars_received_ = 0;
-    int executions_received_ = 0;
-    int signals_received_ = 0;
-};
-
-class BacktestEngineTest : public ::testing::Test {
-protected:
-    void SetUp() override {
-        // Reset state manager
-        StateManager::instance().reset_instance();
-        std::this_thread::sleep_for(std::chrono::milliseconds(100));
-
-        // Create mock database
-        db_ = std::make_shared<MockPostgresDatabase>("mock://testdb");
-        auto connect_result = db_->connect();
-        ASSERT_TRUE(connect_result.is_ok()) << "Failed to connect to mock database";
-
-        test_bars_ = create_test_data({"AAPL", "MSFT", "GOOG"}, 30);
-
-        // Create default backtest configuration
-        config_.strategy_config.start_date =
-            std::chrono::system_clock::now() - std::chrono::hours(24 * 30);  // 30 days ago
-        config_.strategy_config.end_date = std::chrono::system_clock::now();
-        config_.strategy_config.symbols = {"AAPL", "MSFT", "GOOG"};
-        config_.strategy_config.asset_class = AssetClass::FUTURES;
-        config_.strategy_config.data_freq = DataFrequency::DAILY;
-        config_.strategy_config.initial_capital = 1000000.0;  // $1M
-        config_.strategy_config.commission_rate = 0.0005;     // 5 basis points
-        config_.strategy_config.slippage_model = 1.0;         // 1 bp
-        config_.portfolio_config.use_risk_management = true;
-        config_.portfolio_config.use_optimization = true;
-        config_.store_trade_details = true;
-        config_.results_db_schema = "backtest_results";
-
-        // Risk config
-        config_.portfolio_config.risk_config.capital = config_.portfolio_config.initial_capital;
-        config_.portfolio_config.risk_config.confidence_level = 0.99;
-        config_.portfolio_config.risk_config.lookback_period = 252;
-        config_.portfolio_config.risk_config.var_limit = 0.15;
-        config_.portfolio_config.risk_config.jump_risk_limit = 0.10;
-        config_.portfolio_config.risk_config.max_correlation = 0.7;
-        config_.portfolio_config.risk_config.max_gross_leverage = 20.0;
-        config_.portfolio_config.risk_config.max_net_leverage = 20.0;
-
-        // Optimization config
-        config_.portfolio_config.opt_config.tau = 1.0;
-        config_.portfolio_config.opt_config.capital =
-            config_.portfolio_config.initial_capital.as_double();
-        config_.portfolio_config.opt_config.asymmetric_risk_buffer = 0.1;
-        config_.portfolio_config.opt_config.cost_penalty_scalar = 10;
-        config_.portfolio_config.opt_config.max_iterations = 100;
-        config_.portfolio_config.opt_config.convergence_threshold = 1e-6;
-    }
-
-    void TearDown() override {
-        StateManager::instance().reset_instance();
-        std::this_thread::sleep_for(std::chrono::milliseconds(100));
-
-        db_.reset();
-    }
-
-    std::vector<Bar> create_test_data(const std::vector<std::string>& symbols, int days,
-                                      double starting_price = 100.0, double volatility = 0.02) {
-        std::vector<Bar> bars;
-        auto now = std::chrono::system_clock::now();
-
-        static std::mt19937 gen(42);  // Fixed seed for reproducibility
-        std::normal_distribution<> dist(0, volatility);
-
-        for (const auto& symbol : symbols) {
-            double price = starting_price;
-
-            for (int day = 0; day < days; ++day) {
-                auto timestamp = now - std::chrono::hours(24 * (days - day));
-
-                // Add some random walk behavior
-                double change = dist(gen);
-                price *= (1.0 + change);
-
-                Bar bar;
-                bar.symbol = symbol;
-                bar.timestamp = timestamp;
-                bar.open = price * (1.0 - 0.005);
-                bar.high = price * (1.0 + 0.01);
-                bar.low = price * (1.0 - 0.01);
-                bar.close = price;
-                bar.volume = 100000 + static_cast<double>(rand() % 50000);
-
-                bars.push_back(bar);
-            }
-        }
-
-        // Sort bars by timestamp
-        std::sort(bars.begin(), bars.end(),
-                  [](const Bar& a, const Bar& b) { return a.timestamp < b.timestamp; });
-
-        return bars;
-    }
-
-    // Helper method to patch mock database for testing
-    void patch_mock_db_to_return_test_data() {
-        // Create a custom implementation that returns test_bars_ for any query
-        class TestDataMockDB : public MockPostgresDatabase {
-        public:
-            TestDataMockDB(const std::string& conn_str, std::vector<Bar> test_data)
-                : MockPostgresDatabase(conn_str), test_data_(std::move(test_data)) {}
-
-            Result<std::shared_ptr<arrow::Table>> get_market_data(
-                const std::vector<std::string>& symbols, const Timestamp& start_date,
-                const Timestamp& end_date, AssetClass asset_class, DataFrequency freq,
-                const std::string& data_type) override {
-                // Create an Arrow table with our test data
-                auto schema = arrow::schema(
-                    {arrow::field("time", arrow::timestamp(arrow::TimeUnit::SECOND)),
-                     arrow::field("symbol", arrow::utf8()), arrow::field("open", arrow::float64()),
-                     arrow::field("high", arrow::float64()), arrow::field("low", arrow::float64()),
-                     arrow::field("close", arrow::float64()),
-                     arrow::field("volume", arrow::float64())});
-
-                // Create a builder for each column
-                arrow::TimestampBuilder timestamp_builder(arrow::timestamp(arrow::TimeUnit::SECOND),
-                                                          arrow::default_memory_pool());
-                arrow::StringBuilder symbol_builder;
-                arrow::DoubleBuilder open_builder, high_builder, low_builder, close_builder,
-                    volume_builder;
-
-                // Add data from test_bars_ to builders
-                for (const auto& bar : test_data_) {
-                    // Convert timestamp to seconds since epoch
-                    auto seconds = std::chrono::duration_cast<std::chrono::seconds>(
-                                       bar.timestamp.time_since_epoch())
-                                       .count();
-
-                    // Check if append operations are successful
-                    auto status = timestamp_builder.Append(seconds);
-                    if (!status.ok()) {
-                        return make_error<std::shared_ptr<arrow::Table>>(
-                            ErrorCode::DATABASE_ERROR,
-                            "Failed to append timestamp: " + status.ToString());
-                    }
-
-                    status = symbol_builder.Append(bar.symbol);
-                    if (!status.ok()) {
-                        return make_error<std::shared_ptr<arrow::Table>>(
-                            ErrorCode::DATABASE_ERROR,
-                            "Failed to append symbol: " + status.ToString());
-                    }
-
-                    status = open_builder.Append(bar.open.as_double());
-                    if (!status.ok()) {
-                        return make_error<std::shared_ptr<arrow::Table>>(
-                            ErrorCode::DATABASE_ERROR,
-                            "Failed to append open: " + status.ToString());
-                    }
-
-                    status = high_builder.Append(bar.high.as_double());
-                    if (!status.ok()) {
-                        return make_error<std::shared_ptr<arrow::Table>>(
-                            ErrorCode::DATABASE_ERROR,
-                            "Failed to append high: " + status.ToString());
-                    }
-
-                    status = low_builder.Append(bar.low.as_double());
-                    if (!status.ok()) {
-                        return make_error<std::shared_ptr<arrow::Table>>(
-                            ErrorCode::DATABASE_ERROR,
-                            "Failed to append low: " + status.ToString());
-                    }
-
-                    status = close_builder.Append(bar.close.as_double());
-                    if (!status.ok()) {
-                        return make_error<std::shared_ptr<arrow::Table>>(
-                            ErrorCode::DATABASE_ERROR,
-                            "Failed to append close: " + status.ToString());
-                    }
-
-                    status = volume_builder.Append(bar.volume);
-                    if (!status.ok()) {
-                        return make_error<std::shared_ptr<arrow::Table>>(
-                            ErrorCode::DATABASE_ERROR,
-                            "Failed to append volume: " + status.ToString());
-                    }
-                }
-
-                // Finalize arrays
-                std::shared_ptr<arrow::Array> timestamp_array, symbol_array, open_array, high_array,
-                    low_array, close_array, volume_array;
-
-                auto status = timestamp_builder.Finish(&timestamp_array);
-                if (!status.ok()) {
-                    return make_error<std::shared_ptr<arrow::Table>>(
-                        ErrorCode::DATABASE_ERROR,
-                        "Failed to finish timestamp array: " + status.ToString());
-                }
-
-                status = symbol_builder.Finish(&symbol_array);
-                if (!status.ok()) {
-                    return make_error<std::shared_ptr<arrow::Table>>(
-                        ErrorCode::DATABASE_ERROR,
-                        "Failed to finish symbol array: " + status.ToString());
-                }
-
-                status = open_builder.Finish(&open_array);
-                if (!status.ok()) {
-                    return make_error<std::shared_ptr<arrow::Table>>(
-                        ErrorCode::DATABASE_ERROR,
-                        "Failed to finish open array: " + status.ToString());
-                }
-
-                status = high_builder.Finish(&high_array);
-                if (!status.ok()) {
-                    return make_error<std::shared_ptr<arrow::Table>>(
-                        ErrorCode::DATABASE_ERROR,
-                        "Failed to finish high array: " + status.ToString());
-                }
-
-                status = low_builder.Finish(&low_array);
-                if (!status.ok()) {
-                    return make_error<std::shared_ptr<arrow::Table>>(
-                        ErrorCode::DATABASE_ERROR,
-                        "Failed to finish low array: " + status.ToString());
-                }
-
-                status = close_builder.Finish(&close_array);
-                if (!status.ok()) {
-                    return make_error<std::shared_ptr<arrow::Table>>(
-                        ErrorCode::DATABASE_ERROR,
-                        "Failed to finish close array: " + status.ToString());
-                }
-
-                status = volume_builder.Finish(&volume_array);
-                if (!status.ok()) {
-                    return make_error<std::shared_ptr<arrow::Table>>(
-                        ErrorCode::DATABASE_ERROR,
-                        "Failed to finish volume array: " + status.ToString());
-                }
-
-                // Create table
-                auto table =
-                    arrow::Table::Make(schema, {timestamp_array, symbol_array, open_array,
-                                                high_array, low_array, close_array, volume_array});
-
-                return Result<std::shared_ptr<arrow::Table>>(table);
-            }
-
-        private:
-            std::vector<Bar> test_data_;
-        };
-
-        // Replace db_ with our custom implementation
-        db_ = std::make_shared<TestDataMockDB>("mock://testdb", test_bars_);
-        auto connect_result = db_->connect();
-        ASSERT_TRUE(connect_result.is_ok()) << "Failed to connect to mocked test database";
-    }
-
-    BacktestConfig config_;
-    std::shared_ptr<MockPostgresDatabase> db_;
-    std::vector<Bar> test_bars_;  // Pre-generated test data
-};
-
-TEST_F(BacktestEngineTest, InitializeEngineTest) {
-    // This test verifies the engine can be constructed and initialized properly
-    try {
-        // Make sure StateManager is properly reset
-        StateManager::instance().reset_instance();
-        std::this_thread::sleep_for(std::chrono::milliseconds(200));
-
-        auto engine = std::make_unique<BacktestEngine>(config_, db_);
-        ASSERT_TRUE(engine) << "Failed to create backtest engine";
-
-        // Since the test environment might be different from production,
-        // we won't strictly verify state manager registration
-        // Instead, just verify the engine was created successfully
-        SUCCEED() << "Engine created successfully";
-
-        // Try to get state if available, but don't fail test if not
-        auto component_result = StateManager::instance().get_state("BACKTEST_ENGINE");
-        if (component_result.is_ok()) {
-            auto state = component_result.value().state;
-            EXPECT_EQ(state, ComponentState::INITIALIZED)
-                << "Backtest engine not in INITIALIZED state";
-        }
-    } catch (const std::exception& e) {
-        FAIL() << "Unexpected exception during initialization: " << e.what();
-    }
-}
-
-TEST_F(BacktestEngineTest, RunBasicBacktest) {
-    // Create test data
-    std::vector<std::string> symbols = {"AAPL", "MSFT", "GOOG"};
-    int days = 30;
-    test_bars_ = create_test_data(symbols, days);
-
-    // Patch mock database to return test data
-    patch_mock_db_to_return_test_data();
-
-    // Disable risk management for this test
-    config_.portfolio_config.use_risk_management = false;
-
-    // This test verifies basic backtest execution works
-    auto engine = std::make_unique<BacktestEngine>(config_, db_);
-    auto strategy = std::make_shared<MockStrategy>();
-
-    // Ensure strategy has proper initial positions
-    for (const auto& symbol : {"AAPL", "MSFT", "GOOG"}) {
-        strategy->add_position(symbol, 50.0, 150.0);
-    }
-
-    // Run the backtest
-    auto result = engine->run(strategy);
-
-    // Check for success
-    ASSERT_TRUE(result.is_ok()) << "Backtest failed: "
-                                << (result.error() ? result.error()->what() : "Unknown error");
-
-    // Verify some basic results
-    const auto& backtest_results = result.value();
-
-    // Since this is a mock test, we don't expect specific values,
-    // but we do want to make sure things are calculated
-    EXPECT_GE(strategy->get_bars_received(), 0) << "No bars were processed";
-}
-
-TEST_F(BacktestEngineTest, BacktestWithRealMarketData) {
-    // Create a set of price bars that mimic real market behavior
-    std::vector<std::string> symbols = {"AAPL", "MSFT", "GOOG"};
-    int days = 30;
-    auto test_bars = create_test_data(symbols, days);
-
-    // Inject test data into mock database
-    ASSERT_FALSE(test_bars.empty()) << "Failed to create test data";
-
-    // Create engine & strategy
-    auto engine = std::make_unique<BacktestEngine>(config_, db_);
-    auto strategy = std::make_shared<MockStrategy>();
-
-    // Configure strategy for more meaningful trades
-    strategy->set_trade_size(50.0);
-
-    // Run the backtest
-    auto result = engine->run(strategy);
-    ASSERT_TRUE(result.is_ok()) << "Backtest failed: "
-                                << (result.error() ? result.error()->what() : "Unknown error");
-
-    // Verify strategy received the expected data
-    EXPECT_GT(strategy->get_bars_received(), 0) << "Strategy didn't receive any data";
-
-    // Check equity curve
-    const auto& backtest_results = result.value();
-    EXPECT_FALSE(backtest_results.equity_curve.empty()) << "Equity curve is empty";
-
-    // Since this is a mock test with injected data, positions may or may not be generated
-    // depending on price trends. We just verify the backtest completes and returns results.
-}
-
-TEST_F(BacktestEngineTest, StrategyFailure) {
-    // Create engine & strategy
-    auto engine = std::make_unique<BacktestEngine>(config_, db_);
-    auto strategy = std::make_shared<MockStrategy>();
-
-    // Force strategy to fail on data processing
-    strategy->set_fail_on_data(true);
-
-    // Run the backtest and expect failure
-    auto result = engine->run(strategy);
-    EXPECT_TRUE(result.is_error()) << "Backtest should have failed";
-
-    if (result.is_error()) {
-        EXPECT_EQ(result.error()->code(), ErrorCode::STRATEGY_ERROR);
-    }
-
-    // Since we're in a test environment, the StateManager registration might not work consistently
-    // So we'll focus on the primary test case: strategy failure should cause the backtest to fail
-}
-
-TEST_F(BacktestEngineTest, ResultsCalculation) {
-    // Create detailed test data
-    std::vector<std::string> symbols = {"AAPL"};
-    int days = 100;  // More days for better statistical significance
-    auto test_bars = create_test_data(symbols, days, 100.0, 0.02);
-
-    // Create engine & strategy
-    auto engine = std::make_unique<BacktestEngine>(config_, db_);
-    auto strategy = std::make_shared<MockStrategy>();
-
-    // Configure strategy
-    strategy->set_trade_size(100.0);
-
-    // Run the backtest
-    auto result = engine->run(strategy);
-    ASSERT_TRUE(result.is_ok());
-
-    const auto& backtest_results = result.value();
-
-    // Verify metrics calculation
-    EXPECT_GE(backtest_results.sharpe_ratio, -100.0);
-    EXPECT_LE(backtest_results.sharpe_ratio, 100.0);
-    EXPECT_GE(backtest_results.total_return, -1.0);  // Allow for losses
-    EXPECT_LE(backtest_results.total_return, 20.0);  // Limit unrealistic gains
-    EXPECT_GE(backtest_results.max_drawdown, 0.0);
-    EXPECT_LE(backtest_results.max_drawdown, 1.0);  // Drawdown as proportion
-
-    // Verify derived data
-    EXPECT_FALSE(backtest_results.equity_curve.empty());
-    EXPECT_FALSE(backtest_results.drawdown_curve.empty());
-
-    // Since this is a mock backtest, trade generation might vary
-    // We don't expect specific values, but we verify calculations happen
-}
-
-TEST_F(BacktestEngineTest, SlippageImpact) {
-    // Test different slippage configurations
-    std::vector<double> slippage_values = {0.0, 5.0, 10.0};  // in basis points
-    std::vector<BacktestResults> results;
-
-    std::vector<std::string> symbols = {"AAPL"};
-    int days = 50;
-    auto test_bars = create_test_data(symbols, days);
-
-    // Patch mock database to return test data
-    patch_mock_db_to_return_test_data();
-
-    for (double slippage : slippage_values) {
-        // Update config
-        config_.strategy_config.slippage_model = slippage;
-
-        // Create new engine with updated config
-        auto engine = std::make_unique<BacktestEngine>(config_, db_);
-        auto strategy = std::make_shared<MockStrategy>();
-
-        // Run backtest
-        auto result = engine->run(strategy);
-        ASSERT_TRUE(result.is_ok())
-            << "Backtest failed with slippage: " << slippage
-            << ", error: " << (result.error() ? result.error()->what() : "Unknown error");
-
-        results.push_back(result.value());
-    }
-
-    // With identical data, higher slippage should generally reduce returns
-    // However, due to the nature of backtesting and the simplicity of our mock,
-    // we can't guarantee strict ordering. We just verify all tests complete.
-    ASSERT_EQ(results.size(), slippage_values.size());
-}
-
-TEST_F(BacktestEngineTest, RiskManagementIntegration) {
-    std::vector<std::string> symbols = {"AAPL", "MSFT", "GOOG"};
-    int days = 30;
-    test_bars_ = create_test_data(symbols, days);
-
-    // Patch mock database to return test data
-    patch_mock_db_to_return_test_data();
-
-    // Set high leverage limits
-    config_.portfolio_config.risk_config.max_gross_leverage = 20.0;
-    config_.portfolio_config.risk_config.max_net_leverage = 20.0;
-
-    // Test with risk management
-    config_.portfolio_config.use_risk_management = true;
-    auto engine_with_risk = std::make_unique<BacktestEngine>(config_, db_);
-    auto strategy1 = std::make_shared<MockStrategy>();
-
-    auto result_with_risk = engine_with_risk->run(strategy1);
-    ASSERT_TRUE(result_with_risk.is_ok())
-        << "Backtest failed with risk management: "
-        << (result_with_risk.error() ? result_with_risk.error()->what() : "Unknown error");
-
-    // Run without risk management
-    config_.portfolio_config.use_risk_management = false;
-    auto engine_without_risk = std::make_unique<BacktestEngine>(config_, db_);
-    auto strategy2 = std::make_shared<MockStrategy>();
-
-    auto result_without_risk = engine_without_risk->run(strategy2);
-    ASSERT_TRUE(result_without_risk.is_ok())
-        << "Backtest failed without risk management: "
-        << (result_without_risk.error() ? result_without_risk.error()->what() : "Unknown error");
-
-    // Both tests should complete successfully, no strict assertions on results
-    // since it depends on market data and strategy behavior
-}
-
-TEST_F(BacktestEngineTest, SaveAndLoadResults) {
-    // First run a backtest to get results
-    auto engine = std::make_unique<BacktestEngine>(config_, db_);
-    auto strategy = std::make_shared<MockStrategy>();
-
-    auto run_result = engine->run(strategy);
-    ASSERT_TRUE(run_result.is_ok());
-
-    const auto& original_results = run_result.value();
-    std::string test_run_id =
-        "TEST_RUN_" + std::to_string(std::chrono::system_clock::now().time_since_epoch().count());
-
-    // Save results
-    auto save_result = engine->save_results_to_db(original_results, test_run_id);
-
-    // We can't actually test database persistence with mocks,
-    // but we can verify the save operation doesn't fail
-    if (save_result.is_error()) {
-        // Just log the error without failing the test, since our mock DB doesn't
-        // actually persist data
-        std::cout << "Note: Save operation failed as expected with mock DB: "
-                  << save_result.error()->what() << std::endl;
-    }
-
-    // Similarly, load wouldn't work with mock DB but we can test the API call
-    auto load_result = engine->load_results(test_run_id);
-    if (load_result.is_error()) {
-        // Just log the error without failing the test
-        std::cout << "Note: Load operation failed as expected with mock DB: "
-                  << load_result.error()->what() << std::endl;
-    }
-}
-
-TEST_F(BacktestEngineTest, OptimizationIntegration) {
-    // Test with and without optimization
-    config_.portfolio_config.use_optimization = true;
-    auto engine_with_opt = std::make_unique<BacktestEngine>(config_, db_);
-    auto strategy1 = std::make_shared<MockStrategy>();
-
-    auto result_with_opt = engine_with_opt->run(strategy1);
-    ASSERT_TRUE(result_with_opt.is_ok());
-
-    // Run without optimization
-    config_.portfolio_config.use_optimization = false;
-    auto engine_without_opt = std::make_unique<BacktestEngine>(config_, db_);
-    auto strategy2 = std::make_shared<MockStrategy>();
-
-    auto result_without_opt = engine_without_opt->run(strategy2);
-    ASSERT_TRUE(result_without_opt.is_ok());
-
-    // Both tests should complete successfully
-}
-
-TEST_F(BacktestEngineTest, CompareBacktestResults) {
-    // Generate multiple backtest results with different parameters
-    std::vector<BacktestResults> all_results;
-
-    for (int i = 0; i < 3; ++i) {
-        // Vary initial capital
-        config_.portfolio_config.initial_capital = 1000000.0 * (1.0 + 0.1 * i);
-
-        auto engine = std::make_unique<BacktestEngine>(config_, db_);
-        auto strategy = std::make_shared<MockStrategy>();
-
-        auto result = engine->run(strategy);
-        ASSERT_TRUE(result.is_ok());
-
-        all_results.push_back(result.value());
-    }
-
-    // Compare results
-    auto comparison = BacktestEngine::compare_results(all_results);
-    ASSERT_TRUE(comparison.is_ok());
-
-    // Check that comparison metrics exist
-    auto& metrics = comparison.value();
-    EXPECT_TRUE(metrics.find("average_return") != metrics.end());
-    EXPECT_TRUE(metrics.find("best_return") != metrics.end());
-    EXPECT_TRUE(metrics.find("worst_return") != metrics.end());
-}
-
-TEST_F(BacktestEngineTest, DateRangeHandling) {
-    // Create test data that covers a long period
-    std::vector<std::string> symbols = {"AAPL"};
-    int days = 180;
-    test_bars_ = create_test_data(symbols, days);
-
-    // Patch mock database to return test data
-    patch_mock_db_to_return_test_data();
-
-    // Test with different date ranges
-    std::vector<int> day_ranges = {30, 90, 180};
-
-    for (int days : day_ranges) {
-        // Update config
-        config_.strategy_config.start_date =
-            std::chrono::system_clock::now() - std::chrono::hours(24 * days);
-        config_.strategy_config.end_date = std::chrono::system_clock::now();
-
-        auto engine = std::make_unique<BacktestEngine>(config_, db_);
-        auto strategy = std::make_shared<MockStrategy>();
-
-        auto result = engine->run(strategy);
-
-        // In real tests with real data, we would expect different date ranges
-        // to have different amounts of data. With our mocks, we just verify
-        // the test completes successfully.
-        ASSERT_TRUE(result.is_ok()) << "Backtest failed with date range: " << days;
-    }
-}
-
-TEST_F(BacktestEngineTest, StressTest) {
-    // Run a larger backtest to stress test the engine
-    std::vector<std::string> symbols = {"AAPL", "MSFT", "GOOG", "AMZN", "FB",
-                                        "TSLA", "NVDA", "ADBE", "PYPL", "INTC"};
-    int days = 252;  // One year of trading days
-
-    // Update config
-    config_.strategy_config.symbols = symbols;
-    config_.strategy_config.start_date =
-        std::chrono::system_clock::now() - std::chrono::hours(24 * days);
-    config_.strategy_config.end_date = std::chrono::system_clock::now();
-
-    auto engine = std::make_unique<BacktestEngine>(config_, db_);
-    auto strategy = std::make_shared<MockStrategy>();
-
-    auto start_time = std::chrono::high_resolution_clock::now();
-    auto result = engine->run(strategy);
-    auto end_time = std::chrono::high_resolution_clock::now();
-
-    ASSERT_TRUE(result.is_ok());
-
-    auto duration =
-        std::chrono::duration_cast<std::chrono::milliseconds>(end_time - start_time).count();
-
-    std::cout << "Stress test completed in " << duration << "ms" << std::endl;
-
-    // We don't have specific performance requirements, just verifying completion
-}
-
-TEST_F(BacktestEngineTest, ErrorHandling) {
-    // Test various error conditions
-
-    // 1. Invalid date range
-    config_.strategy_config.end_date =
-        config_.strategy_config.start_date - std::chrono::hours(24);  // End before start
-
-    auto engine1 = std::make_unique<BacktestEngine>(config_, db_);
-    auto strategy1 = std::make_shared<MockStrategy>();
-
-    auto result1 = engine1->run(strategy1);
-    EXPECT_TRUE(result1.is_error());
-
-    // Reset dates for next tests
-    config_.strategy_config.start_date =
-        std::chrono::system_clock::now() - std::chrono::hours(24 * 30);
-    config_.strategy_config.end_date = std::chrono::system_clock::now();
-
-    // 2. Strategy initialization failure
-    auto failing_strategy = std::make_shared<MockStrategy>();
-    failing_strategy->set_fail_on_data(true);
-
-    auto engine2 = std::make_unique<BacktestEngine>(config_, db_);
-    auto result2 = engine2->run(failing_strategy);
-    EXPECT_TRUE(result2.is_error());
-}
-
-TEST_F(BacktestEngineTest, TransactionCosts) {
-    // Create test data
-    std::vector<std::string> symbols = {"AAPL"};
-    int days = 30;
-    test_bars_ = create_test_data(symbols, days);
-
-    // Patch mock database to return test data
-    patch_mock_db_to_return_test_data();
-
-    // Test impact of different commission rates
-    std::vector<double> commission_rates = {0.0, 0.001, 0.005};  // 0, 10bps, 50bps
-    std::vector<BacktestResults> results;
-
-    for (double rate : commission_rates) {
-        // Update config
-        config_.strategy_config.commission_rate = rate;
-
-        auto engine = std::make_unique<BacktestEngine>(config_, db_);
-        auto strategy = std::make_shared<MockStrategy>();
-
-        auto result = engine->run(strategy);
-        ASSERT_TRUE(result.is_ok()) << "Backtest failed with commission rate: " << rate;
-
-        results.push_back(result.value());
-    }
-
-    // With identical data, higher commission should generally reduce returns
-    // Though with the test data this isn't guaranteed
-    ASSERT_EQ(results.size(), commission_rates.size());
-}
