+++ tests/core/test_equity_portfolio_namespace.cpp
