+++ tests/live/corp_actions/test_effect_classes.cpp
