+++ tests/strategy/test_pnl_accounting_branch.cpp
